% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_GAMM.R
\name{simul_modGAMM}
\alias{simul_modGAMM}
\title{Function to simulate y in GAMM()  model}
\usage{
simul_modGAMM(seed = 1)
}
\arguments{
\item{seed}{Seed used for simulation}
}
\value{
expected difference
}
\description{
Function to simulate y in GAMM() model
}
\examples{
simul_modlogist(seed = 1)

}
