% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_GAMM.R, R/simulate_logist.R
\name{simul_modlogist}
\alias{simul_modlogist}
\title{Function to simulate y in GAMM()  model}
\usage{
simul_modlogist(seed = 1)

simul_modlogist(seed = 1)
}
\arguments{
\item{seed}{Seed used for simulation}

\item{mod}{Give the structure of the model}
}
\value{
expected difference

expected difference
}
\description{
Function to simulate y in GAMM() model

Function to simulate y in logistf() linear model
}
\examples{
simul_modlogist(seed = 1)

simul_modlogist(seed = 1)

}
