% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_LO.R
\docType{data}
\name{theme_LO}
\alias{theme_LO}
\title{Theme for ggplot}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 13.
}
\usage{
theme_LO
}
\value{

}
\description{
Object: Sober theme for ggplot
}
\examples{
PLOT_MAIN_Text<-ggplot + theme_LO
}
\keyword{datasets}
